[Enhancement][Olana] Navbar UI Change
