+++ app/(protected)/(tabs)/catalogue.tsx
