+++ app/(protected)/view-post/_layout.tsx
