+++ app/(protected)/view-post/full-screen/[index].tsx
