--- app/(tabs)/index.tsx
@@ -1,9 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function Index() {
-    return (
-        <View>
-            <Text>Index</Text>
-        </View>
-    );
-}
