+++ app/index.tsx
@@ -1,5 +1,5 @@
 import { Redirect } from "expo-router";
 
 export default function Root() {
-    return <Redirect href="/(tabs)/home" />;
+    return <Redirect href="/(protected)/(tabs)/home" />;
 }
