+++ babel.config.js
@@ -5,6 +5,5 @@ module.exports = function (api) {
             ["babel-preset-expo", { jsxImportSource: "nativewind" }],
             "nativewind/babel",
         ],
-        plugins: ["react-native-reanimated/plugin"],
     };
 };
