+++ metro.config.js
@@ -1,11 +1,8 @@
 const { getDefaultConfig } = require("expo/metro-config");
 const { withNativeWind } = require("nativewind/metro");
-const {
-    wrapWithReanimatedMetroConfig,
-} = require("react-native-reanimated/metro-config");
 
 const config = getDefaultConfig(__dirname);
 
-module.exports = withNativeWind(wrapWithReanimatedMetroConfig(config), {
+module.exports = withNativeWind(config, {
     input: "./global.css",
 });
